# web/fixtures/openapi.json
{
  "openapi": "3.1.0",
  "info": { "title": "SentinelTrace API", "version": "1.0.0" },
  "servers": [{ "url": "/api/v1" }],
  "paths": {
...
    "/users": {
      "get": {
        "operationId": "listUsers",
        "parameters": [
          { "name": "q", "in": "query", "schema": { "type": "string" } },
          { "name": "department", "in": "query", "schema": { "type": "string" } },
          { "name": "role", "in": "query", "schema": { "type": "string" } },
          { "name": "min_risk", "in": "query", "schema": { "type": "number" } },
          { "name": "departing_within", "in": "query", "schema": { "type": "integer" } },
          { "name": "sort", "in": "query", "schema": { "type": "string" } },
          { "name": "limit", "in": "query", "schema": { "type": "integer" } },
          { "name": "cursor", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "OK", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UserListResponse" } } } } }
      }
    },
    "/users/{user_id}": {
      "get": {
        "operationId": "getUser",
        "parameters": [{ "name": "user_id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": {
          "200": { "description": "OK", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UserProfile" } } } },
          "404": { "description": "Not found", "content": { "application/problem+json": { "schema": { "$ref": "#/components/schemas/ProblemDetails" } } } }
        }
      }
    },
    "/users/{user_id}/risk": {
      "get": {
        "operationId": "getUserRisk",
        "parameters": [
          { "name": "user_id", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "date_from", "in": "query", "schema": { "type": "string", "format": "date" } },
          { "name": "date_to", "in": "query", "schema": { "type": "string", "format": "date" } },
          { "name": "include", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "OK", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UserRiskSeries" } } } } }
      }
    },
    "/users/{user_id}/timeline": {
      "get": {
        "operationId": "getUserTimeline",
        "parameters": [
          { "name": "user_id", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "date", "in": "query", "required": true, "schema": { "type": "string", "format": "date" } }
        ],
        "responses": { "200": { "description": "OK", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UserTimeline" } } } } }
      }
    },
    "/incidents": {
      "get": {
        "operationId": "listIncidents",
        "parameters": [
          { "name": "lane", "in": "query", "schema": { "type": "string" } },
          { "name": "status", "in": "query", "schema": { "type": "string" } },
          { "name": "min_risk", "in": "query", "schema": { "type": "number" } },
          { "name": "min_confidence", "in": "query", "schema": { "type": "number" } },
          { "name": "user_id", "in": "query", "schema": { "type": "string" } },
          { "name": "date_from", "in": "query", "schema": { "type": "string", "format": "date" } },
          { "name": "date_to", "in": "query", "schema": { "type": "string", "format": "date" } },
          { "name": "stage_max", "in": "query", "schema": { "type": "integer" } },
          { "name": "campaign_id", "in": "query", "schema": { "type": "string" } },
          { "name": "sort", "in": "query", "schema": { "type": "string" } },
          { "name": "limit", "in": "query", "schema": { "type": "integer" } },
          { "name": "cursor", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "OK", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/IncidentListResponse" } } } } }
      }
    },
    "/incidents/{incident_id}": {
      "get": {
        "operationId": "getIncident",
        "parameters": [{ "name": "incident_id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": {
          "200": { "description": "OK", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/IncidentDetail" } } } },
          "404": { "description": "Not found", "content": { "application/problem+json": { "schema": { "$ref": "#/components/schemas/ProblemDetails" } } } }
        }
      }
    },
    "/incidents/{incident_id}/graph": {
      "get": {
        "operationId": "getIncidentGraph",
        "parameters": [{ "name": "incident_id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "OK", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/IncidentGraph" } } } } }
      }
    },
    "/incidents/{incident_id}/export": {
      "get": {
        "operationId": "exportIncident",
        "parameters": [
          { "name": "incident_id", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "format", "in": "query", "schema": { "type": "string", "enum": ["json", "pdf"] } }
        ],
        "responses": { "200": { "description": "OK", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/EvidencePack" } } } } }
      }
    },
    "/incidents/{incident_id}/review": {
      "post": {
        "operationId": "reviewIncident",
        "parameters": [
          { "name": "incident_id", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "force", "in": "query", "schema": { "type": "boolean" } }
        ],
        "requestBody": { "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ReviewRequest" } } } },
        "responses": {
          "201": { "description": "Created", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ReviewResponse" } } } },
          "409": { "description": "Conflict", "content": { "application/problem+json": { "schema": { "$ref": "#/components/schemas/ProblemDetails" } } } }
        }
      }
    },
    "/campaigns": {
      "get": {
        "operationId": "listCampaigns",
        "parameters": [
          { "name": "user_id", "in": "query", "schema": { "type": "string" } },
          { "name": "min_stage", "in": "query", "schema": { "type": "integer" } },
          { "name": "limit", "in": "query", "schema": { "type": "integer" } }
        ],
        "responses": { "200": { "description": "OK", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CampaignListResponse" } } } } }
      }
    },
    "/campaigns/{campaign_id}": {
      "get": {
        "operationId": "getCampaign",
        "parameters": [{ "name": "campaign_id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "OK", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Campaign" } } } } }
      }
    },
...
    "/rules/{rule_id}/stats": {
      "get": {
        "operationId": "getRuleStats",
        "parameters": [
          { "name": "rule_id", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "date_from", "in": "query", "schema": { "type": "string", "format": "date" } },
          { "name": "date_to", "in": "query", "schema": { "type": "string", "format": "date" } }
        ],
        "responses": { "200": { "description": "OK", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/RuleStats" } } } } }
      }
    },
...
    "/eval/report": {
      "get": {
        "operationId": "getEvalReport",
        "parameters": [{ "name": "config_version", "in": "query", "schema": { "type": "string" } }],
        "responses": { "200": { "description": "OK", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/EvalReport" } } } } }
      }
    },
    "/analyze": {
      "post": {
        "operationId": "analyze",
        "requestBody": { "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AnalyzeRequest" } } } },
        "responses": { "200": { "description": "OK", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AnalyzeResponse" } } } } }
      }
    },
    "/auth/login": {
      "post": {
        "operationId": "login",
        "requestBody": { "content": { "application/json": { "schema": { "$ref": "#/components/schemas/LoginRequest" } } } },
        "responses": {
          "200": { "description": "OK", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/LoginResponse" } } } },
          "401": { "description": "Unauthorized", "content": { "application/problem+json": { "schema": { "$ref": "#/components/schemas/ProblemDetails" } } } }
        }
      }
    },
...
    "/auth/me": {
      "get": {
        "operationId": "getMe",
        "responses": {
          "200": { "description": "OK", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SessionUser" } } } },
          "401": { "description": "Unauthorized", "content": { "application/problem+json": { "schema": { "$ref": "#/components/schemas/ProblemDetails" } } } }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": { "type": "string" },
          "title": { "type": "string" },
          "status": { "type": "integer" },
          "detail": { "type": "string" },
          "instance": { "type": "string" },
          "trace_id": { "type": "string" }
        },
        "required": ["type", "title", "status"]
      },
...
      "UserListResponse": {
        "type": "object",
        "properties": {
          "items": { "type": "array", "items": { "$ref": "#/components/schemas/UserListItem" } },
          "next_cursor": { "type": ["string", "null"] },
          "total": { "type": "integer" }
        }
      },
      "UserProfile": {
        "type": "object",
        "properties": {
          "user_id": { "type": "string" },
          "name": { "type": "string" },
          "email": { "type": "string" },
          "org": {
            "type": "object",
            "properties": {
              "role": { "type": "string" },
              "department": { "type": "string" },
              "team": { "type": "string" },
              "supervisor": { "type": "string" },
              "cohort_size": { "type": "integer" }
            }
          },
          "tenure_days": { "type": "integer" },
          "departure_date": { "type": ["string", "null"] },
          "first_seen": { "type": "string" },
          "last_seen": { "type": "string" },
          "baseline": {
            "type": "object",
            "properties": {
              "days_available": { "type": "integer" },
              "maturity": { "type": "number" },
              "working_window": {
                "type": "object",
                "properties": { "start_min": { "type": "integer" }, "end_min": { "type": "integer" } }
              }
            }
          },
          "current_risk": { "type": "number" },
          "risk_ewma": { "type": "number" },
          "incident_counts": {
            "type": "object",
            "properties": {
              "AUTO_FLAG": { "type": "integer" },
              "ANALYST_REVIEW": { "type": "integer" },
              "MONITOR": { "type": "integer" }
            }
          }
        }
      },
      "UserRiskSeries": {
        "type": "object",
        "properties": {
          "user_id": { "type": "string" },
          "window": { "type": "object", "properties": { "from": { "type": "string" }, "to": { "type": "string" } } },
          "series": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "date": { "type": "string" },
                "risk": { "type": "number" },
                "confidence": { "type": "number" },
                "risk_ewma": { "type": "number" },
                "signal_count": { "type": "integer" },
                "incident_ids": { "type": "array", "items": { "type": "string" } }
              }
            }
          },
          "peer_band": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "date": { "type": "string" },
                "p50": { "type": "number" },
                "p90": { "type": "number" },
                "p99": { "type": "number" }
              }
            }
          },
          "summary": {
            "type": "object",
            "properties": {
              "peak_risk": { "type": "number" },
              "peak_date": { "type": "string" },
              "days_above_threshold": { "type": "integer" },
              "trend": { "type": "string" }
            }
          }
        }
      },
...
      "UserTimeline": {
        "type": "object",
        "properties": {
          "date": { "type": "string" },
          "events": { "type": "array", "items": { "$ref": "#/components/schemas/TimelineEvent" } },
          "working_window": {
            "type": "object",
            "properties": { "start_min": { "type": "integer" }, "end_min": { "type": "integer" } }
          },
          "final_risk": { "type": "number" }
        }
      },
...
      "IncidentListResponse": {
        "type": "object",
        "properties": {
          "items": { "type": "array", "items": { "$ref": "#/components/schemas/IncidentListItem" } },
          "next_cursor": { "type": ["string", "null"] },
          "total": { "type": "integer" },
          "facets": {
            "type": "object",
            "properties": {
              "lane": { "type": "object", "additionalProperties": { "type": "integer" } },
              "max_stage": { "type": "object", "additionalProperties": { "type": "integer" } }
            }
          }
        }
      },
...
      "IncidentDetail": {
        "type": "object",
        "properties": {
          "incident_id": { "type": "string" },
          "user": {
            "type": "object",
            "properties": {
              "user_id": { "type": "string" },
              "name": { "type": "string" },
              "role": { "type": "string" },
              "department": { "type": "string" },
              "cohort_size": { "type": "integer" }
            }
          },
          "window": {
            "type": "object",
            "properties": { "start": { "type": "string" }, "end": { "type": "string" }, "duration_min": { "type": "number" } }
          },
          "score": {
            "type": "object",
            "properties": {
              "risk": { "type": "number" },
              "confidence": { "type": "number" },
              "triage_lane": { "$ref": "#/components/schemas/TriageLane" },
              "breakdown": {
                "type": "object",
                "properties": {
                  "prior_logit": { "type": "number" },
                  "rule_points": { "type": "number" },
                  "ml_points": { "type": "number" },
                  "correlation_points": { "type": "number" },
                  "total_logit": { "type": "number" },
                  "tau": { "type": "number" }
                }
              },
              "confidence_terms": {
                "type": "object",
                "properties": {
                  "agreement": { "type": "number" },
                  "diversity": { "type": "number" },
                  "completeness": { "type": "number" },
                  "maturity": { "type": "number" },
                  "caps_applied": { "type": "array", "items": { "type": "string" } }
                }
              }
            }
          },
          "narrative": {
            "type": "object",
            "properties": {
              "headline": { "type": "string" },
              "summary": { "type": "string" },
              "detail_bullets": { "type": "array", "items": { "type": "string" } },
              "template_ids": { "type": "array", "items": { "type": "string" } }
            }
          },
          "signals": { "type": "array", "items": { "$ref": "#/components/schemas/Signal" } },
          "attribution": {
            "type": "object",
            "properties": {
              "note": { "type": "string" },
              "items": { "type": "array", "items": { "$ref": "#/components/schemas/AttributionItem" } },
              "minimal_sufficient_set": { "type": "array", "items": { "type": "string" } },
              "alert_threshold": { "type": "number" }
            }
          },
          "campaign": {
            "type": ["object", "null"],
            "properties": {
              "campaign_id": { "type": "string" },
              "incident_count": { "type": "integer" },
              "first_seen": { "type": "string" },
              "last_seen": { "type": "string" },
              "stage_progression": { "type": "array", "items": { "type": "integer" } }
            }
          },
          "status": { "type": "string", "enum": ["open", "closed"] },
          "review": {
            "type": ["object", "null"],
            "properties": {
              "verdict": { "type": "string" },
              "note": { "type": "string" },
              "analyst_id": { "type": "string" },
              "reviewed_at": { "type": "string" }
            }
          },
          "config_version": { "type": "string" },
          "links": {
            "type": "object",
            "properties": { "graph": { "type": "string" }, "export": { "type": "string" } }
          }
        }
      },
...
      "IncidentGraph": {
        "type": "object",
        "properties": {
          "incident_id": { "type": "string" },
          "over_dense": { "type": "boolean" },
          "nodes": { "type": "array", "items": { "$ref": "#/components/schemas/GraphNode" } },
          "edges": { "type": "array", "items": { "$ref": "#/components/schemas/GraphEdge" } },
          "layout_hint": { "type": "string" },
          "stats": {
            "type": "object",
            "properties": {
              "node_count": { "type": "integer" },
              "edge_count": { "type": "integer" },
              "component_diameter": { "type": "integer" }
            }
          }
        }
      },
      "EvidencePack": {
        "type": "object",
        "properties": {
          "incident": { "$ref": "#/components/schemas/IncidentDetail" },
          "raw_events": { "type": "array", "items": { "type": "object", "additionalProperties": true } },
          "config_snapshot": { "type": "object", "additionalProperties": true },
          "watermark": {
            "type": "object",
            "properties": { "exported_by": { "type": "string" }, "exported_at": { "type": "string" } }
          }
        }
      },
...
      "ReviewRequest": {
        "type": "object",
        "properties": {
          "verdict": { "type": "string", "enum": ["confirmed_threat", "benign", "inconclusive"] },
          "note": { "type": "string" },
          "analyst_id": { "type": "string" },
          "time_to_triage_sec": { "type": "integer" },
          "propose_suppression": { "oneOf": [{ "$ref": "#/components/schemas/SuppressionProposal" }, { "type": "null" }] }
        },
        "required": ["verdict", "analyst_id"]
      },
      "ReviewResponse": {
        "type": "object",
        "properties": {
          "review_id": { "type": "integer" },
          "incident_id": { "type": "string" },
          "verdict": { "type": "string" },
          "reviewed_at": { "type": "string" },
          "incident_status": { "type": "string" },
          "suppression": { "type": ["object", "null"], "properties": { "status": { "type": "string" } } },
          "effects": {
            "type": "object",
            "properties": {
              "rule_stats_updated": { "type": "array", "items": { "type": "string" } },
              "user_risk_ewma_adjusted": { "type": "boolean" }
            }
          }
        }
      },
...
      "Campaign": {
        "type": "object",
        "properties": {
          "campaign_id": { "type": "string" },
          "user_id": { "type": "string" },
          "first_seen": { "type": "string" },
          "last_seen": { "type": "string" },
          "incident_count": { "type": "integer" },
          "peak_risk": { "type": "number" },
          "campaign_risk": { "type": "number" },
          "max_stage": { "type": "integer" },
          "stage_progression": { "type": "array", "items": { "$ref": "#/components/schemas/CampaignStageEvent" } },
          "narrative": { "type": "string" }
        }
      },
...
      "CampaignListResponse": {
        "type": "object",
        "properties": {
          "items": { "type": "array", "items": { "$ref": "#/components/schemas/CampaignListItem" } },
          "total": { "type": "integer" }
        }
      },
...
      "RuleStats": {
        "type": "object",
        "properties": {
          "rule_id": { "type": "string" },
          "configured_weight": { "type": "number" },
          "fire_count": { "type": "integer" },
          "fire_rate_per_user_day": { "type": "number" },
          "reviewed": { "type": "integer" },
          "confirmed": { "type": "integer" },
          "benign": { "type": "integer" },
          "inconclusive": { "type": "integer" },
          "observed_precision": { "type": "number" },
          "measured_log_odds": { "type": "number" },
          "weight_drift": { "type": "number" },
          "recommendation": { "type": "string" }
        }
      },
...
      "EvalReport": {
        "type": "object",
        "properties": {
          "config_version": { "type": "string" },
          "data_source": { "type": "string" },
          "generated_at": { "type": "string" },
          "primary_metrics": {
            "type": "object",
            "properties": {
              "insider_recall": { "type": "array", "items": { "type": "number" } },
              "incident_precision": { "type": "array", "items": { "type": "number" } },
              "auto_flag_precision": { "type": "array", "items": { "type": "number" } },
              "pr_auc": { "type": "array", "items": { "type": "number" } },
              "median_time_to_detect_days": { "type": "number" }
            }
          },
          "pr_curve": {
            "type": "array",
            "items": { "type": "object", "properties": { "precision": { "type": "number" }, "recall": { "type": "number" } } }
          },
          "per_scenario_recall": { "type": "object", "additionalProperties": { "type": "number" } },
          "time_to_detect_distribution": { "type": "array", "items": { "type": "number" } }
        }
      },
...
    }
  }
}